Proof of concept.
Sends the pen over rather than the whole canvas image data (unusably slow). Need to be VERY careful of losing the session, if the dashboard reloads we've lost the image, this is probably the only performant way to do this so far, even sending just the dirty area is pretty rough above reasonable canvas size. Could maybe compress the image data (libpng?) before sending then decompressing on the dashboard side, or store the pen events in Node then resync that over?